src/app_reservations/views.py
from rest_framework import viewsets
from rest_framework.response import Response
from rest_framework import status
from .models import Reservation
from .serializers import ReservationSerializer
from .utils import send_to_telegram


class ReservationViewSet(viewsets.ModelViewSet):
    queryset = Reservation.objects.all()
    serializer_class = ReservationSerializer

    def perform_create(self, serializer):
        reservation = serializer.save()
        send_to_telegram(reservation.name, reservation.phone, reservation.created_at)
